feat: coordinatorlayout, dialog 수정
