--- app/src/main/java/com/example/lecturesopt28th/base/BaseFragment.kt
@@ -1,29 +1,0 @@
-package com.example.lecturesopt28th.base
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.databinding.ViewDataBinding
-import androidx.fragment.app.Fragment
-
-abstract class BaseFragment <T: ViewDataBinding>: Fragment() {
-    private var _binding:T? = null
-    protected val binding get() = _binding ?: error("view binding error")
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        _binding = getFragmentBinding(inflater, container)
-        return binding.root
-    }
-
-    abstract fun getFragmentBinding(inflater: LayoutInflater, container: ViewGroup?): T
-
-    override fun onDestroy() {
-        super.onDestroy()
-        _binding = null
-    }
-}
