+++ app/src/main/java/com/example/lecturesopt28th/di/DataSourceModule.kt
@@ -6,10 +6,10 @@ import com.example.lecturesopt28th.githubrepo.data.source.GithubRepoDataSourceIm
 import com.example.lecturesopt28th.home.api.SearchUserApiService
 import com.example.lecturesopt28th.home.data.source.SearchUserDataSource
 import com.example.lecturesopt28th.home.data.source.SearchUserDataSourceImpl
-import com.example.lecturesopt28th.login.LoginApiService
-import com.example.lecturesopt28th.login.LoginDataSource
-import com.example.lecturesopt28th.signup.SignUpApiService
-import com.example.lecturesopt28th.signup.SignUpDataSource
+import com.example.lecturesopt28th.login.api.LoginApiService
+import com.example.lecturesopt28th.login.data.source.LoginDataSource
+import com.example.lecturesopt28th.signup.api.SignUpApiService
+import com.example.lecturesopt28th.signup.data.source.SignUpDataSource
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
