+++ app/src/main/java/com/example/lecturesopt28th/di/RepositoryModule.kt
@@ -6,12 +6,12 @@ import com.example.lecturesopt28th.githubrepo.data.repository.GithubRepoReposito
 import com.example.lecturesopt28th.home.data.source.SearchUserDataSource
 import com.example.lecturesopt28th.home.data.repository.SearchUserRepository
 import com.example.lecturesopt28th.home.data.repository.SearchUserRepositoryImpl
-import com.example.lecturesopt28th.login.LoginDataSource
-import com.example.lecturesopt28th.login.LoginRepository
-import com.example.lecturesopt28th.login.LoginRepositoryImpl
-import com.example.lecturesopt28th.signup.SignUpDataSource
-import com.example.lecturesopt28th.signup.SignUpRepositoryImpl
-import com.example.lecturesopt28th.signup.SignUpRespository
+import com.example.lecturesopt28th.login.data.source.LoginDataSource
+import com.example.lecturesopt28th.login.repository.LoginRepository
+import com.example.lecturesopt28th.login.repository.LoginRepositoryImpl
+import com.example.lecturesopt28th.signup.data.source.SignUpDataSource
+import com.example.lecturesopt28th.signup.repository.SignUpRepositoryImpl
+import com.example.lecturesopt28th.signup.repository.SignUpRespository
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
