--- app/src/main/java/com/example/lecturesopt28th/login/LoginApiService.kt
@@ -1,10 +1,0 @@
-package com.example.lecturesopt28th.login
-
-import io.reactivex.Single
-import retrofit2.http.Body
-import retrofit2.http.POST
-
-interface LoginApiService {
-    @POST("login/signin")
-    fun login(@Body body: RequestLogin): Single<ResponseLogin>
-}
