--- app/src/main/java/com/example/lecturesopt28th/login/LoginRepository.kt
@@ -1,7 +1,0 @@
-package com.example.lecturesopt28th.login
-
-import io.reactivex.Single
-
-interface LoginRepository {
-    fun login(requestLogin: RequestLogin): Single<ResponseLogin>
-}
