--- app/src/main/java/com/example/lecturesopt28th/login/LoginRepositoryImpl.kt
@@ -1,11 +1,0 @@
-package com.example.lecturesopt28th.login
-
-import io.reactivex.Single
-import javax.inject.Inject
-
-class LoginRepositoryImpl @Inject constructor(
-    private val dataSource: LoginDataSource
-): LoginRepository {
-    override fun login(requestLogin: RequestLogin): Single<ResponseLogin> =
-        dataSource.login(requestLogin)
-}
