--- app/src/main/java/com/example/lecturesopt28th/login/RequestLogin.kt
@@ -1,6 +1,0 @@
-package com.example.lecturesopt28th.login
-
-data class RequestLogin(
-    val email: String?,
-    val password: String?
-)
