+++ app/src/main/java/com/example/lecturesopt28th/login/data/dto/RequestLogin.kt
@@ -1,0 +1,6 @@
+package com.example.lecturesopt28th.login.data.dto
+
+data class RequestLogin(
+    val email: String?,
+    val password: String?
+)
