+++ app/src/main/java/com/example/lecturesopt28th/login/repository/LoginRepository.kt
@@ -1,0 +1,9 @@
+package com.example.lecturesopt28th.login.repository
+
+import com.example.lecturesopt28th.login.data.dto.RequestLogin
+import com.example.lecturesopt28th.login.data.dto.ResponseLogin
+import io.reactivex.Single
+
+interface LoginRepository {
+    fun login(requestLogin: RequestLogin): Single<ResponseLogin>
+}
