--- app/src/main/java/com/example/lecturesopt28th/signup/SignUpRespository.kt
@@ -1,7 +1,0 @@
-package com.example.lecturesopt28th.signup
-
-import io.reactivex.Single
-
-interface SignUpRespository {
-    fun signUp(requestSignUp: RequestSignUp): Single<ResponseSingUp>
-}
