+++ app/src/main/java/com/example/lecturesopt28th/signup/data/dto/ResponseSingUp.kt
@@ -1,0 +1,11 @@
+package com.example.lecturesopt28th.signup.data.dto
+
+data class ResponseSingUp(
+    val success: Boolean,
+    val message: String,
+    val data: Data
+) {
+    data class Data(
+        val nickname: String
+    )
+}
