+++ app/src/main/java/com/example/lecturesopt28th/utils/InputChecker.kt
@@ -1,0 +1,10 @@
+package com.example.lecturesopt28th.utils
+
+import android.widget.EditText
+import androidx.lifecycle.LiveData
+
+object InputChecker {
+    fun checkBlank(listText: List<LiveData<String>>): Int {
+        return listText.filter { it.value.isNullOrEmpty() }.count()
+    }
+}
